+++ app/app/utils.py
@@ -6,6 +6,8 @@ import time
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.contrib.gis.geoip2 import GeoIP2
+from django.db.models import Lookup
+from django.db.models.fields import Field
 from django.utils import timezone
 
 import requests
@@ -18,6 +20,20 @@ from marketing.utils import get_or_save_email_subscriber
 from social_django.models import UserSocialAuth
 
 logger = logging.getLogger(__name__)
+
+
+@Field.register_lookup
+class NotEqual(Lookup):
+    """Allow lookup and exclusion using not equal."""
+
+    lookup_name = 'ne'
+
+    def as_sql(self, compiler, connection):
+        """Handle as SQL method for not equal lookup."""
+        lhs, lhs_params = self.process_lhs(compiler, connection)
+        rhs, rhs_params = self.process_rhs(compiler, connection)
+        params = lhs_params + rhs_params
+        return f'%s <> %s' % (lhs, rhs), params
 
 
 def ellipses(data, _len=75):
